requirements.txt
fastapi
pydantic
uvicorn[standard]
sqlalchemy
psycopg2-binary
redis
pymongo
pyjwt
python-jose[cryptography] 
passlib[bcrypt]
python-multipart
pandas
python-dotenv